+++ src/lib/pages/upload/index.tsx
@@ -1,59 +1,32 @@
-import { Heading, Text } from "@chakra-ui/react";
-import { useRouter } from "next/router";
-import { useEffect } from "react";
+import { useState } from "react";
 
-import {
-  useCurrentChain,
-  useInternalNavigate,
-  useWasmConfig,
-} from "lib/app-provider";
-import { ConnectWalletAlert } from "lib/components/ConnectWalletAlert";
-import { Stepper } from "lib/components/stepper";
-import { UploadSection } from "lib/components/upload/UploadSection";
-import WasmPageContainer from "lib/components/WasmPageContainer";
-import { AmpEvent, AmpTrack } from "lib/services/amplitude";
-import { useUploadAccessParams } from "lib/services/proposalService";
-import type { HumanAddr } from "lib/types";
-import { AccessConfigPermission } from "lib/types";
+import type { UploadTxInternalResult } from "lib/app-provider";
+import { useWasmConfig } from "lib/app-provider";
 
-const Upload = () => {
-  const router = useRouter();
-  const { address } = useCurrentChain();
-  const navigate = useInternalNavigate();
-  const { data, isLoading } = useUploadAccessParams();
+import { UploadComplete } from "./completed";
+import { Upload } from "./upload";
 
-  const isPermissionedNetwork =
-    data?.permission !== AccessConfigPermission.EVERYBODY;
-
-  const enableUpload =
-    !isPermissionedNetwork ||
-    Boolean(data?.addresses?.includes(address as HumanAddr));
-
+const UploadIndex = () => {
   useWasmConfig({ shouldRedirect: true });
+  const [txInfo, setTxInfo] = useState<UploadTxInternalResult>({
+    codeDisplayName: "",
+    codeId: "",
+    txHash: "",
+    formattedFee: "",
+  });
 
-  useEffect(() => {
-    // Redirect back to deploy page
-    if (!enableUpload && !isLoading)
-      navigate({ pathname: "/deploy", replace: true });
-    else if (router.isReady) AmpTrack(AmpEvent.TO_UPLOAD);
-  }, [enableUpload, isLoading, navigate, router.isReady]);
+  const [completed, setCompleted] = useState(false);
 
-  return (
-    <WasmPageContainer>
-      <Text variant="body1" color="text.dark" mb={3} fontWeight={700}>
-        DEPLOY NEW CONTRACT
-      </Text>
-      <Stepper mode="deploy" currentStep={1} />
-      <Heading as="h5" variant="h5" my={12}>
-        Upload Wasm file
-      </Heading>
-      <ConnectWalletAlert
-        subtitle="You need to connect your wallet first"
-        mb={12}
-      />
-      <UploadSection handleBack={() => router.back()} />
-    </WasmPageContainer>
+  return completed ? (
+    <UploadComplete txResult={txInfo} />
+  ) : (
+    <Upload
+      onComplete={(txResult: UploadTxInternalResult) => {
+        setTxInfo(txResult);
+        setCompleted(true);
+      }}
+    />
   );
 };
 
-export default Upload;
+export default UploadIndex;